pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "flow-engine"
version = "0.1.0"
description = "Declarative flow orchestration engine with Starlark and Pydantic V2"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pydantic>=2.5",
  "PyYAML>=6",
  "starlark-pyo3>=2024.0",
  "jsonschema>=4.20.0",
]

[project.optional-dependencies]
dev = ["pytest>=7", "pytest-asyncio>=0.23"]
mysql = [
  "sqlalchemy>=2.0",
  "pymysql>=1.1",
  "python-dotenv>=1.0",
  "mcp>=1.2.0",
]
api = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.32",
  "openpyxl>=3.1",
  "python-multipart>=0.0.9",
]
runner = [
  "croniter>=2.0",
]

[project.scripts]
flow-api = "flow_engine.http_api:main"
flow-db = "flow_engine.db.cli:main"
flow-worker = "flow_engine.runner.worker:main"
flow-coordinator = "flow_engine.runner.coordinator:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.setuptools.packages.find]
where = ["src"]
include = ["flow_engine*"]

[tool.setuptools.package-data]
flow_engine = ["starlib/registry.json", "starlib/internal/**/*.star"]
